==> notebooks/02_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv("../data/processed/interactions.csv")

data.head()

data.isna().sum()

data.duplicated().sum()

data = data.drop_duplicates(subset=["user_id", "item_id", "timestamp"])

data.duplicated().sum()

user_counts = data.user_id.value_counts()
item_counts = data.item_id.value_counts()

active_users = user_counts[user_counts >= 5].index
popular_items = item_counts[item_counts >= 5].index

print(len(active_users))
print(len(popular_items))

interactions = data[
    data.user_id.isin(active_users) &
    data.item_id.isin(popular_items)
]

interactions["interaction"] = 1

interactions = interactions[[
    "user_id", "item_id", "interaction", "timestamp"
]]

interactions.head()

### Split interaction data into test and train based on Time

interactions = interactions.sort_values("timestamp")
split_time = interactions.timestamp.quantile(0.8)

train = interactions[interactions.timestamp <= split_time]
test = interactions[interactions.timestamp > split_time]

print("Train interactions:", len(train))
print("Test interactions:", len(test))

print("Users in train:", train.user_id.nunique())
print("Users in test:", test.user_id.nunique())

train.to_csv("../data/processed/train.csv", index=False)
test.to_csv("../data/processed/test.csv", index=False)
