==> notebooks/03_baseline_models.ipynb <==
import pandas as pd
import numpy as np

train = pd.read_csv("../data/processed/train.csv")
test = pd.read_csv("../data/processed/test.csv")

train.head()

test.head()

### Popularity Baseline : Find top 10 popular items

top_items = (
    train.groupby("item_id")
    .size()
    .sort_values(ascending=False)
    .head(10)
    .index
    .tolist()
)

top_items

### Evaluate Popularity Model (Recall@K)

def recall_at_k(test, recommendations, k=10):
    recalls = []

    for user in test.user_id.unique():
        true_items = set(
            test[test.user_id == user].item_id
        )

        if not true_items:
            continue

        recs = set(recommendations[:k])
        recall = len(true_items & recs) / len(true_items)
        recalls.append(recall)

    return sum(recalls) / len(recalls)

pop_recall = recall_at_k(test, top_items, k=10)

pop_recall

### User-Based Collaborative Filtering

user_item = (
    train.pivot_table(
        index="user_id",
        columns="item_id",
        values="interaction",
        fill_value=0
    )
)

user_item.head()

from sklearn.metrics.pairwise import cosine_similarity

user_sim = cosine_similarity(user_item)

### Evalute User-based model (Recall@K)

def recommend_user_cf(user_idx, user_item, user_sim, k=10):
    sim_scores = user_sim[user_idx]
    weighted_scores = sim_scores @ user_item.values
    item_indices = np.argsort(weighted_scores)[::-1][:k]
    return user_item.columns[item_indices].tolist()

user_index = {u: i for i, u in enumerate(user_item.index)}
user_index

recalls = []

for user in test.user_id.unique():
    if user not in user_index:
        continue

    idx = user_index[user]
    recs = recommend_user_cf(idx, user_item, user_sim, k=10)

    true_items = set(
        test[test.user_id == user].item_id
    )

    recall = len(true_items & set(recs)) / len(true_items)
    recalls.append(recall)

user_cf_recall = sum(recalls) / len(recalls)
user_cf_recall

import pickle
