==> notebooks/06_final_demo.ipynb <==
import sys
import os

PROJECT_ROOT = os.path.abspath(os.path.join(os.getcwd(), ".."))
sys.path.append(PROJECT_ROOT)

from src.inference.recommend import MFRecommender
import pickle

user_map = pickle.load(open("../src/models/user_map.pkl", "rb"))
item_map = pickle.load(open("../src/models/item_map.pkl", "rb"))
reverse_item_map = pickle.load(open("../src/models/reverse_item_map.pkl", "rb"))

type(user_map), type(item_map)

recommender = MFRecommender(
    "../src/models/mf_model.pkl",
    user_map,
    item_map,
    reverse_item_map
)
